+++ packages/postcss-config/src/index.ts
@@ -1,74 +1,50 @@
-import { deepMerge, isObject } from '@minko-fe/lodash-pro'
+import { isObject, merge } from '@minko-fe/lodash-pro'
 import { type PxtoremOptions } from '@minko-fe/postcss-pxtorem'
 import { type PxtoviewportOptions } from '@minko-fe/postcss-pxtoviewport'
 import { createRequire } from 'node:module'
-import { type AcceptedPlugin } from 'postcss'
 import { type pluginOptions } from 'postcss-preset-env'
 import { type Config } from 'tailwindcss'
 
 const _require = createRequire(import.meta.url)
 
-export type PostcssConfig =
-  | {
-      /**
-       * @default true
-       * @description set false in vite env
-       * @description default built-in
-       */
-      'postcss-import'?: boolean
-      /**
-       * @default true
-       * @description default built-in
-       */
-      'tailwindcss/nesting'?: boolean
-
-      /**
-       * @default true
-       */
-      'autoprefixer'?: boolean
-      /**
-       * @default false
-       */
-      'postcss-pxtorem'?: false | PxtoremOptions
-      /**
-       * @default false
-       */
-      'postcss-pxtoviewport'?: false | PxtoviewportOptions
-      /**
-       * @default true
-       * @description default built-in
-       */
-      'tailwindcss'?: string | Config | { config: string | Config } | boolean
-      /**
-       * @default true
-       * @description default built-in
-       */
-      'postcss-preset-env'?: pluginOptions | boolean
-    }
-  | undefined
-
-const defaultOptions: Required<PostcssConfig> = {
-  'postcss-import': true,
-  'tailwindcss/nesting': true,
-  'tailwindcss': true,
-  'postcss-pxtorem': false,
-  'postcss-pxtoviewport': false,
-  'autoprefixer': true,
-  'postcss-preset-env': true,
+export type PostcssConfig = {
+  /**
+   * @default true
+   * @description set false in vite env
+   * @description default built-in
+   */
+  'postcss-import'?: boolean
+  /**
+   * @default true
+   * @description default built-in
+   */
+  'tailwindcss/nesting'?: boolean
+  /**
+   * @default true
+   */
+  'autoprefixer'?: boolean
+  /**
+   * @default false
+   */
+  'postcss-pxtorem'?: false | PxtoremOptions
+  /**
+   * @default false
+   */
+  'postcss-pxtoviewport'?: false | PxtoviewportOptions
+  /**
+   * @default true
+   * @description default built-in
+   */
+  'tailwindcss'?: string | Config | { config: string | Config } | boolean
+  /**
+   * @default false
+   * @description default built-in
+   * useful for nextjs
+   */
+  'postcss-preset-env'?: pluginOptions | boolean
 }
 
-const plugins: AcceptedPlugin[] = []
-
-const pluginsForNextjs: (string | any[])[] = []
-
-function unPlugins(pluginName: string, options?: any) {
-  plugins.push(options ? _require(pluginName)(options) : _require(pluginName))
-  pluginsForNextjs.push(options ? [pluginName, options] : pluginName)
-}
-
-const postcssConfig = (options: PostcssConfig) => {
-  options = deepMerge(defaultOptions, options || {})!
-
+export const postcssConfig = (options: PostcssConfig, unPlugins: (pluginName: string, opts?: any) => void) => {
   if (options['postcss-import'] !== false) {
     unPlugins('postcss-import')
   }
@@ -130,22 +106,14 @@ const postcssConfig = (options: PostcssConfig) => {
         stage: 3,
       }
 
-      const _options = deepMerge(defaultOptions, isObject(presetEnv) ? presetEnv : {})
+      const _options = merge(defaultOptions, isObject(presetEnv) ? presetEnv : {})
 
       if (options['tailwindcss/nesting'] === false) {
+        _options.features ??= {}
         _options.features['nesting-rules'] = true
       }
 
       unPlugins('postcss-preset-env', _options)
     }
   }
-
-  return {
-    normal: plugins,
-    nextjs: pluginsForNextjs,
-  }
-}
-
-export function definePlugins(options: PostcssConfig) {
-  return postcssConfig(options)
 }