+++ packages/postcss-config/src/tailwindcss/plugin/index.ts
@@ -1,1 +1,0 @@
-export { center } from './center'