+++ packages/vite-config/src/node/utils/index.ts
@@ -1,6 +1,3 @@
-import { getTsconfig } from 'get-tsconfig'
-import path from 'node:path'
-
 export function injectEnv(envConf: Record<string, any>): ImportMetaEnv {
   const ret: any = {}
 
@@ -17,43 +14,4 @@ export function injectEnv(envConf: Record<string, any>): ImportMetaEnv {
   }
 
   return ret
-}
-
-export interface PathMapping {
-  find: RegExp | string
-  replacement: string
-}
-
-export function pathsMapToAlias(root: string) {
-  const tsconfig = getTsconfig(root)
-
-  const compilerOptions = tsconfig?.config.compilerOptions
-
-  function resolvePathMappings(paths: Record<string, string[]>, base: string) {
-    const sortedPatterns = Object.keys(paths).sort((a: string, b: string) => getPrefixLength(b) - getPrefixLength(a))
-    const resolved: PathMapping[] = []
-    for (let pattern of sortedPatterns) {
-      const relativePaths = paths[pattern]
-      if (relativePaths.length > 1) {
-        continue
-      }
-      pattern = escapeStringRegexp(pattern).replaceAll('/*', '')
-      resolved.push({
-        find: pattern,
-        replacement: path.resolve(base, relativePaths[0].replaceAll('/*', '')),
-      })
-    }
-
-    return resolved
-  }
-  function getPrefixLength(pattern: string): number {
-    const prefixLength = pattern.indexOf('*')
-    return pattern.slice(0, Math.max(0, prefixLength)).length
-  }
-
-  function escapeStringRegexp(string: string) {
-    return string.replaceAll(/[$()+.?[\\\]^{|}]/g, '\\$&').replaceAll('-', '\\x2d')
-  }
-
-  return resolvePathMappings(compilerOptions?.paths || {}, compilerOptions?.baseUrl || '.')
 }