+++ playground/spa/postcss.config.cjs
@@ -1,5 +1,0 @@
-const { definePlugins } = require('@minko-fe/postcss-config')
-
-module.exports = {
-  plugins: definePlugins().normal,
-}